+++ src/fides/config/admin_ui_settings.py
@@ -1,3 +1,4 @@
+from enum import Enum
 from typing import Optional
 
 from pydantic import Field, SerializeAsAny
@@ -6,6 +7,11 @@ from pydantic_settings import SettingsConfigDict
 from fides.api.custom_types import AnyHttpUrlStringRemovesSlash
 
 from .fides_settings import FidesSettings
+
+
+class ErrorNotificationMode(str, Enum):
+    CONSOLE_ONLY = "console_only"
+    TOAST = "toast"
 
 
 class AdminUISettings(FidesSettings):
@@ -20,5 +26,9 @@ class AdminUISettings(FidesSettings):
     max_privacy_request_download_rows: int = Field(
         default=100000,
         description="The maximum number of rows permitted to be returned in a privacy request report download",
+    )
+    error_notification_mode: Optional[str] = Field(
+        default=ErrorNotificationMode.CONSOLE_ONLY.value,
+        description="This setting controls how errors are notified to users.",
     )
     model_config = SettingsConfigDict(env_prefix="FIDES__ADMIN_UI__")